feat(discard)!: implement
Add support for Clojure's `#_`
a.k.a. [discarding](https://clojure.org/guides/weird_characters#_discard)

It supports stacking and ignoring trailing/overstacking,
including when discarding top-level forms.

Example can be seen in the tests (next commit).

This could be deemed a breaking change since
we no longer interpret `#_` as `(hashfn _)`.

It works by adding a new prefix to the parser, to be more specific -
the logic for parsing prefixes is refactored to allow for early
detection of the `#_` prefix (and possibly more two-character prefixes)
and more cleanly handle the rest of prefixes.
It keeps track of the number of stacked discards on each depth, applying
them when `dispatch`-ing. The current depth depends on the calls to
`open-table` and `close-table`.